requirements.txt
langchain
langchain-community
langchain-ollama
chromadb
python-dotenv
unstructured[md]